Address https://github.com/umccr/orcabus/pull/27/files#r1230409633 by @mmalenic... not entirely happy with the refactor since ApiDoc also had to move over security.rs, can be improved
