+++ skel/rust-api/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod db;
 pub mod file;
+pub mod security;
